refactor: move MainActivity and ViewModel to ui package
