+++ app/src/main/java/dev/enesky/moviecatalog/ui/MainViewModel.kt
@@ -1,4 +1,4 @@
-package dev.enesky.moviecatalog
+package dev.enesky.moviecatalog.ui
 
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
